Added implementation of the PairUp operator.
PairUp combines pairs of adjacent elements in the source sequence and returns the result of applying a user-defined projection to them. PairUp can operate on either odd-length or even-length sequences. In the case of add-length sequences, an optional imbalance strategy can be supplied, which defines whether PairUp should skip the last (unpaired) value, or process the terminal element with a default value pairing. The result of PairUp applied to a non-empty sequence, will always be a sequence of half the length of the source sequence.