Added implementation of the Slice operator.
Slice is an optimized, clear version of s.Skip(start).Take(count). It provides a means to extract the middle portion of a sequence starting at a given index, and extracting up to a specified number of elements. Slice has a performance optimization for indexable sequence that implement IList<T> that allows it to stream just those elements without iterating over the entire sequence.