Added implementation of the Lead operator.
Lead projects a sequence by combining each element with a subsequent element at a given offset. For elements less than the lead offset away from the end, the lead value is considered to be default(T). An overload exists that allows the caller to supply an alternate default value. Lead behaves in a deferred and streaming manner.