No need to Batch to stream batch elements; closes #200
There is no need for Batch to stream elements of each batch, which seems to
be there really to prevent the caller from down-casting to the actual type
returned. However, it leads to unnecessary performance issues when used with
other LINQ operators optimized for lists and collections (like ElementAt or
Count).