Better seeding for Random
Currently it creates its randomness source using the default
constructor of the Random class. But the well known problem with that
is that this is seeded with the system time, which typically doesn't
change for 16ms. So all instances of the random class created during
that interval will return the same sequence.