Remove ToDelimitedString overloads without delimiter arg
The overloads have been obsolete since version 2.1.

`ToDelimitedString` overloads accepting zero arguments and using the
list separator from the current culture is of very marginal utility.
The results can be surprising as they can tend to vary across cultures
and systems when writing cross-platform code. It is best to require
the caller to supply the delimiter explicitly and it's the caller's
choice to use the list separator string from the current thread's
culture, which is what the zero-arg overloads do.