Make package copyright more manageable
This puts the copyright notice extenral to main project file and
friendlier to eyes and version control management (fewer chances of
merge conflicts). It also takes care of an annoying problem where
Visual Studio would re-encoding the `&#169;` to the copyright
symbol whenever the project file would be opened in the IDE and
which would show up as artificial changes ready for commit to the
version control repository.