Fix ordered await bug; closes #436
There were two nasty bugs. First, there was a problem with capturing
of a modified closure in a for-loop causing the ordering keys to be
garbage. Second, an exit condition was missing when looping and
releasing withheld results.

This commit is wording revision of the commit message associated with
7cb39df625089d74c92a811538fe594a83653373 and contains no other code
changes since.