Add conflict-free extensions/generator
With extension methods, there's always a potential for conflict with
some future version of .NET or another library. When the `MoreLinq`
namespace is imported, the user gets all extensions when only a handful
may be needed. What would be helpful instead is to be able to choose
exactly the operators one would like to import. We therefore add a code
generator that is now responsible for generating conflict-free wrappers
to the extension methods at build-time (when built from the command-line
using the build scripts). The wrappers are designed to be used with the
static imports feature of C# introduced with version 6 of the language,
like so:

    using static MoreLinq.Extensions.LagExtension;
    using static MoreLinq.Extensions.LeadExtension;
    
Note that `LagExtension` and `LeadExtension` here are classes and
each hosts its own extension method.

The generator uses Roslyn to parse the project sources and discover the
public extension methods on `MoreEnumerable` (minus those marked
obsolete). For each method found, a class named after the method, and
with the suffix `Extension`, is added to host a forwarding wrapper
extension method. For example, for the method `Pairwise`, the generated
class is named `PairwiseExtension` and which is placed into the
`MoreLinq.Extensions` namespace. An extension method with the same
signature, attributes and documentation as the original is then added
to the class, and repeated for each overload. The implementation of the
wrapper extension method simply calls the original method, passing all
arguments without any changes.

This is a merge of PR #235 that closes #214